Balanced Array
Given an array of even size N, the task is to find the minimum value that can be added to an element so that the array becomes balanced. An array is balanced if the sum of the left half of the array elements is equal to the sum of the right half.